requirements.txt
langchain-huggingface
langchain-community
faiss-cpu
transformers
torch
numpy
pandas
tqdm
